+++ python/tvm/ir_builder.py
@@ -15,7 +15,9 @@
 # specific language governing permissions and limitations
 # under the License.
 """Developer API of IR node builder make function."""
-from __future__ import absolute_import as _abs
+from tvm.runtime import ObjectGeneric, DataType
+
+from ._ffi.base import string_types
 
 from . import api as _api
 from . import stmt as _stmt
@@ -23,9 +25,6 @@ from . import expr as _expr
 from . import make as _make
 from . import ir_pass as _pass
 from . import container as _container
-from ._ffi.base import string_types
-from ._ffi.object_generic import ObjectGeneric
-from ._ffi.runtime_ctypes import TVMType
 from .expr import Call as _Call
 
 class WithScope(object):
@@ -78,7 +77,7 @@ class BufferVar(ObjectGeneric):
         return self._content_type
 
     def __getitem__(self, index):
-        t = TVMType(self._content_type)
+        t = DataType(self._content_type)
         if t.lanes > 1:
             index = _make.Ramp(index * t.lanes, 1, t.lanes)
         return _make.Load(self._content_type, self._buffer_var, index)
@@ -89,7 +88,7 @@ class BufferVar(ObjectGeneric):
             raise ValueError(
                 "data type does not match content type %s vs %s" % (
                     value.dtype, self._content_type))
-        t = TVMType(self._content_type)
+        t = DataType(self._content_type)
         if t.lanes > 1:
             index = _make.Ramp(index * t.lanes, 1, t.lanes)
         self._builder.emit(_make.Store(self._buffer_var, value, index))
