+++ python/tvm/rpc/client.py
@@ -22,12 +22,12 @@ import socket
 import struct
 import time
 import tvm._ffi
+from tvm.contrib import util
+from tvm._ffi.base import TVMError
+from tvm.runtime import ndarray as nd
+from tvm.runtime import load_module as _load_module
 
 from . import base
-from ..contrib import util
-from .._ffi.base import TVMError
-from .._ffi import ndarray as nd
-from ..module import load as _load_module
 
 
 class RPCSession(object):
