+++ python/tvm/rpc/server.py
@@ -38,10 +38,10 @@ import sys
 import signal
 import tvm._ffi
 
-from .._ffi.base import py_str
-from .._ffi.libinfo import find_lib_path
-from ..module import load as _load_module
-from ..contrib import util
+from tvm._ffi.base import py_str
+from tvm._ffi.libinfo import find_lib_path
+from tvm.runtime.module import load as _load_module
+from tvm.contrib import util
 from . import base
 from . base import TrackerCode
