+++ python/tvm/runtime/module.py
@@ -14,22 +14,99 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Container of compiled functions of TVM."""
-from __future__ import absolute_import as _abs
 
+# pylint: disable=invalid-name, unused-import
+"""Runtime Module namespace."""
+import ctypes
 import struct
 from collections import namedtuple
+
 import tvm._ffi
+from tvm._ffi.base import _LIB, check_call, c_str, string_types, _RUNTIME_ONLY
+from tvm._ffi.libinfo import find_include_path
+from .packed_func import PackedFunc, PackedFuncHandle, _set_class_module
 
-from ._ffi.module import ModuleBase, _set_class_module
-from ._ffi.libinfo import find_include_path
-from .contrib import cc as _cc, tar as _tar, util as _util
 
+# profile result of time evaluator
 ProfileResult = namedtuple("ProfileResult", ["mean", "results"])
 
 
-class Module(ModuleBase):
-    """Module container of all TVM generated functions"""
+class Module(object):
+    """Runtime Module."""
+    __slots__ = ["handle", "_entry", "entry_name"]
+
+    def __init__(self, handle):
+        self.handle = handle
+        self._entry = None
+        self.entry_name = "__tvm_main__"
+
+    def __del__(self):
+        check_call(_LIB.TVMModFree(self.handle))
+
+    def __hash__(self):
+        return ctypes.cast(self.handle, ctypes.c_void_p).value
+
+    @property
+    def entry_func(self):
+        """Get the entry function
+
+        Returns
+        -------
+        f : Function
+            The entry function if exist
+        """
+        if self._entry:
+            return self._entry
+        self._entry = self.get_function(self.entry_name)
+        return self._entry
+
+    def get_function(self, name, query_imports=False):
+        """Get function from the module.
+
+        Parameters
+        ----------
+        name : str
+            The name of the function
+
+        query_imports : bool
+            Whether also query modules imported by this module.
+
+        Returns
+        -------
+        f : Function
+            The result function.
+        """
+        ret_handle = PackedFuncHandle()
+        check_call(_LIB.TVMModGetFunction(
+            self.handle, c_str(name),
+            ctypes.c_int(query_imports),
+            ctypes.byref(ret_handle)))
+        if not ret_handle.value:
+            raise AttributeError(
+                "Module has no function '%s'" %  name)
+        return PackedFunc(ret_handle, False)
+
+    def import_module(self, module):
+        """Add module to the import list of current one.
+
+        Parameters
+        ----------
+        module : Module
+            The other module.
+        """
+        check_call(_LIB.TVMModImport(self.handle, module.handle))
+
+    def __getitem__(self, name):
+        if not isinstance(name, string_types):
+            raise ValueError("Can only take string as function name")
+        return self.get_function(name)
+
+    def __call__(self, *args):
+        if self._entry:
+            return self._entry(*args)
+        # pylint: disable=not-callable
+        return self.entry_func(*args)
+
 
     def __repr__(self):
         return "Module(%s, %x)" % (self.type_key, self.handle.value)
@@ -84,93 +161,6 @@ class Module(ModuleBase):
         Module.export_library : export the module to shared library.
         """
         _SaveToFile(self, file_name, fmt)
-
-    def export_library(self,
-                       file_name,
-                       fcompile=None,
-                       **kwargs):
-        """Export the module and its imported device code one library.
-
-        This function only works on host llvm modules.
-        It will pack all the imported modules
-
-        Parameters
-        ----------
-        file_name : str
-            The name of the shared library.
-
-        fcompile : function(target, file_list, kwargs), optional
-            Compilation function to use create dynamic library.
-            If fcompile has attribute object_format, will compile host library
-            to that format. Otherwise, will use default format "o".
-
-        kwargs : dict, optional
-            Additional arguments passed to fcompile
-        """
-        from pathlib import Path
-        if isinstance(file_name, Path):
-            file_name = str(file_name)
-
-        if self.type_key == "stackvm":
-            if not file_name.endswith(".stackvm"):
-                raise ValueError("Module[%s]: can only be saved as stackvm format."
-                                 "did you build with LLVM enabled?" % self.type_key)
-            self.save(file_name)
-            return
-
-        modules = self._collect_dso_modules()
-        temp = _util.tempdir()
-        files = []
-        is_system_lib = False
-        has_c_module = False
-        llvm_target_triple = None
-        for index, module in enumerate(modules):
-            if fcompile is not None and hasattr(fcompile, "object_format"):
-                object_format = fcompile.object_format
-            else:
-                if module.type_key == "llvm":
-                    object_format = "o"
-                else:
-                    assert module.type_key == "c"
-                    object_format = "cc"
-                    has_c_module = True
-            path_obj = temp.relpath("lib" + str(index) + "." + object_format)
-            module.save(path_obj)
-            files.append(path_obj)
-            is_system_lib = (module.type_key == "llvm" and
-                             module.get_function("__tvm_is_system_module")())
-            llvm_target_triple = (module.type_key == "llvm" and
-                                  module.get_function("_get_target_triple")())
-        if not fcompile:
-            if file_name.endswith(".tar"):
-                fcompile = _tar.tar
-            else:
-                fcompile = _cc.create_shared
-
-        if llvm_target_triple is None and hasattr(fcompile, "get_target_triple"):
-            llvm_target_triple = fcompile.get_target_triple()
-
-        if self.imported_modules:
-            if enabled("llvm") and llvm_target_triple:
-                path_obj = temp.relpath("devc.o")
-                m = _PackImportsToLLVM(self, is_system_lib, llvm_target_triple)
-                m.save(path_obj)
-                files.append(path_obj)
-            else:
-                path_cc = temp.relpath("devc.cc")
-                with open(path_cc, "w") as f:
-                    f.write(_PackImportsToC(self, is_system_lib))
-                files.append(path_cc)
-
-        if has_c_module:
-            options = []
-            if "options" in kwargs:
-                opts = kwargs["options"]
-                options = opts if isinstance(opts, (list, tuple)) else [opts]
-            opts = options + ["-I" + path for path in find_include_path()]
-            kwargs.update({'options': opts})
-
-        fcompile(file_name, files, **kwargs)
 
     def time_evaluator(self, func_name, ctx, number=10, repeat=1, min_repeat_ms=0):
         """Get an evaluator that measures time cost of running function.
@@ -249,6 +239,100 @@ class Module(ModuleBase):
     def _dso_exportable(self):
         return self.type_key == "llvm" or self.type_key == "c"
 
+    def export_library(self,
+                       file_name,
+                       fcompile=None,
+                       **kwargs):
+        """Export the module and its imported device code one library.
+
+        This function only works on host llvm modules.
+        It will pack all the imported modules
+
+        Parameters
+        ----------
+        file_name : str
+            The name of the shared library.
+
+        fcompile : function(target, file_list, kwargs), optional
+            Compilation function to use create dynamic library.
+            If fcompile has attribute object_format, will compile host library
+            to that format. Otherwise, will use default format "o".
+
+        kwargs : dict, optional
+            Additional arguments passed to fcompile
+        """
+        # NOTE: this function depends on contrib library features
+        # which are only available in when TVM function is available.
+        if _RUNTIME_ONLY:
+            raise RuntimeError("Cannot call export_library in runtime only mode")
+        # Extra dependencies during runtime.
+        from pathlib import Path
+        from tvm.contrib import cc as _cc, tar as _tar, util as _util
+
+        if isinstance(file_name, Path):
+            file_name = str(file_name)
+
+        if self.type_key == "stackvm":
+            if not file_name.endswith(".stackvm"):
+                raise ValueError("Module[%s]: can only be saved as stackvm format."
+                                 "did you build with LLVM enabled?" % self.type_key)
+            self.save(file_name)
+            return
+
+        modules = self._collect_dso_modules()
+        temp = _util.tempdir()
+        files = []
+        is_system_lib = False
+        has_c_module = False
+        llvm_target_triple = None
+        for index, module in enumerate(modules):
+            if fcompile is not None and hasattr(fcompile, "object_format"):
+                object_format = fcompile.object_format
+            else:
+                if module.type_key == "llvm":
+                    object_format = "o"
+                else:
+                    assert module.type_key == "c"
+                    object_format = "cc"
+                    has_c_module = True
+            path_obj = temp.relpath("lib" + str(index) + "." + object_format)
+            module.save(path_obj)
+            files.append(path_obj)
+            is_system_lib = (module.type_key == "llvm" and
+                             module.get_function("__tvm_is_system_module")())
+            llvm_target_triple = (module.type_key == "llvm" and
+                                  module.get_function("_get_target_triple")())
+        if not fcompile:
+            if file_name.endswith(".tar"):
+                fcompile = _tar.tar
+            else:
+                fcompile = _cc.create_shared
+
+        if llvm_target_triple is None and hasattr(fcompile, "get_target_triple"):
+            llvm_target_triple = fcompile.get_target_triple()
+
+        if self.imported_modules:
+            if enabled("llvm") and llvm_target_triple:
+                path_obj = temp.relpath("devc.o")
+                m = _PackImportsToLLVM(self, is_system_lib, llvm_target_triple)
+                m.save(path_obj)
+                files.append(path_obj)
+            else:
+                path_cc = temp.relpath("devc.cc")
+                with open(path_cc, "w") as f:
+                    f.write(_PackImportsToC(self, is_system_lib))
+                files.append(path_cc)
+
+        if has_c_module:
+            options = []
+            if "options" in kwargs:
+                opts = kwargs["options"]
+                options = opts if isinstance(opts, (list, tuple)) else [opts]
+            opts = options + ["-I" + path for path in find_include_path()]
+            kwargs.update({'options': opts})
+
+        fcompile(file_name, files, **kwargs)
+
 
 def system_lib():
     """Get system-wide library module singleton.
@@ -296,9 +380,13 @@ def load(path, fmt=""):
     # High level handling for .o and .tar file.
     # We support this to be consistent with RPC module load.
     if path.endswith(".o"):
+        # Extra dependencies during runtime.
+        from tvm.contrib import cc as _cc
         _cc.create_shared(path + ".so", path)
         path += ".so"
     elif path.endswith(".tar"):
+        # Extra dependencies during runtime.
+        from tvm.contrib import cc as _cc, util as _util, tar as _tar
         tar_temp = _util.tempdir(custom_path=path.replace('.tar', ''))
         _tar.untar(path, tar_temp.temp_dir)
         files = [tar_temp.relpath(x) for x in tar_temp.listdir()]
@@ -333,5 +421,6 @@ def enabled(target):
     return _Enabled(target)
 
 
-tvm._ffi._init_api("tvm.module")
 _set_class_module(Module)
+
+tvm._ffi._init_api("tvm.module", "tvm.runtime.module")
